Split passkey authentication button and dialog
